added Stub for ParameterisationDependentMeasurementResultTest and adapted the other tests by removing the methods which will be tested in the newly created Test Class.
