+++ Core/src/test/java/de/uka/ipd/sdq/beagle/core/measurement/LoopRepetitionCountMeasurementResultTest.java
@@ -42,54 +42,4 @@ public class LoopRepetitionCountMeasurementResultTest {
 		fail("Not yet implemented");
 	}
 
-	/**
-	 * Test method for
-	 * {@link ParameterisationDependentMeasurementResult#ParameterisationDependentMeasurementResult()}
-	 * .
-	 */
-	@Test
-	public void testParameterisationDependentMeasurementResult() {
-		fail("Not yet implemented");
-	}
-
-	/**
-	 * Test method for
-	 * {@link ParameterisationDependentMeasurementResult#ParameterisationDependentMeasurementResult(Parameterisation)}
-	 * .
-	 */
-	@Test
-	public void testParameterisationDependentMeasurementResultParameterisation() {
-		fail("Not yet implemented");
-	}
-
-	/**
-	 * Test method for
-	 * {@link ParameterisationDependentMeasurementResult#getParameterisation()}
-	 * .
-	 */
-	@Test
-	public void testGetParameterisation() {
-		fail("Not yet implemented");
-	}
-
-	/**
-	 * Test method for
-	 * {@link ParameterisationDependentMeasurementResult#equals(java.lang.Object)}
-	 * and {@link ParameterisationDependentMeasurementResult#hashCode()}.
-	 */
-	@Test
-	public void testEqualsAndHashCode() {
-		fail("Not yet implemented");
-	}
-
-	/**
-	 * Test method for
-	 * {@link ParameterisationDependentMeasurementResult#toString()}
-	 * .
-	 */
-	@Test
-	public void testToString() {
-		fail("Not yet implemented");
-	}
-
 }
