+++ Core/src/test/java/de/uka/ipd/sdq/beagle/core/measurement/ParameterisationDependentMeasurementResultTest.java
@@ -1,0 +1,63 @@
+package de.uka.ipd.sdq.beagle.core.measurement;
+
+import static org.junit.Assert.fail;
+
+import org.junit.Test;
+
+/**
+ * Tests {@link ParameterisationDependentMeasurementResult} and contains all test cases needed
+ * to check every method.
+ * 
+ * @author Annika Berger
+ */
+public class ParameterisationDependentMeasurementResultTest {
+
+
+	/**
+	 * Test method for
+	 * {@link ParameterisationDependentMeasurementResult#ParameterisationDependentMeasurementResult()}
+	 * .
+	 */
+	@Test
+	public void testParameterisationDependentMeasurementResult() {
+		fail("Not yet implemented");
+	}
+
+	/**
+	 * Test method for
+	 * {@link ParameterisationDependentMeasurementResult#ParameterisationDependentMeasurementResult(Parameterisation)}
+	 * .
+	 */
+	@Test
+	public void testParameterisationDependentMeasurementResultParameterisation() {
+		fail("Not yet implemented");
+	}
+
+	/**
+	 * Test method for
+	 * {@link ParameterisationDependentMeasurementResult#getParameterisation()} .
+	 */
+	@Test
+	public void testGetParameterisation() {
+		fail("Not yet implemented");
+	}
+
+	/**
+	 * Test method for
+	 * {@link ParameterisationDependentMeasurementResult#equals(java.lang.Object)} and
+	 * {@link ParameterisationDependentMeasurementResult#hashCode()}.
+	 */
+	@Test
+	public void testEqualsObject() {
+		fail("Not yet implemented");
+	}
+
+	/**
+	 * Test method for {@link ParameterisationDependentMeasurementResult#toString()} .
+	 */
+	@Test
+	public void testToString() {
+		fail("Not yet implemented");
+	}
+
+}
